Turn on C++11 support for GCC and Clang

+++ src/3rdParty/salomesmesh/src/SMESH/SMESH_MeshEditor.cpp
@@ -6017,7 +6017,7 @@ SMESH_MeshEditor::Sew_Error
   // sew the border to the side 2
   // ============================
 
-  int nbNodes[]  = { nSide[0].size(), nSide[1].size() };
+  int nbNodes[]  = { (int)nSide[0].size(), (int)nSide[1].size() };
   int maxNbNodes = Max( nbNodes[0], nbNodes[1] );
 
   TListOfListOfNodes nodeGroupsToMerge;
